Personal Projects/BES - Bulk Email System/requirements.txt
cs50
Flask
Flask-Session
requests
python-docx
PyPDF2
opencv-python
pytesseract
xlrd
cryptography
flask-paginate
urllib3==1.26.7
python-dateutil
pandas
